user-service/requirements.txt
grpcio==1.27.1
grpcio-tools==1.27.1
protobuf==3.11.3
psycopg2==2.8.4
python-dotenv==0.11.0
six==1.14.0
SQLAlchemy==1.3.13
